refactor into new file structure
